Improve code structure and re-usability
