+++ lib/helpers/style_utils.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+
+const mySendButtonTextStyle = TextStyle(
+  color: Colors.lightBlueAccent,
+  fontWeight: FontWeight.bold,
+  fontSize: 18.0,
+);
+
+const myMessageTextFieldDecoration = InputDecoration(
+  contentPadding: EdgeInsets.symmetric(vertical: 10.0, horizontal: 20.0),
+  hintText: 'Type your message here...',
+  border: InputBorder.none,
+);
+
+const myMessageContainerDecoration = BoxDecoration(
+  border: Border(
+    top: BorderSide(color: Colors.lightBlueAccent, width: 2.0),
+  ),
+);
+const myTextFieldDecoration = InputDecoration(
+hintText: 'Enter information',
+contentPadding:
+EdgeInsets.symmetric(vertical: 10.0, horizontal: 20.0),
+border: OutlineInputBorder(
+borderRadius: BorderRadius.all(Radius.circular(32.0)),
+),
+enabledBorder: OutlineInputBorder(
+borderSide: BorderSide(color: Colors.blueAccent, width: 1.0),
+borderRadius: BorderRadius.all(Radius.circular(32.0)),
+),
+focusedBorder: OutlineInputBorder(
+borderSide: BorderSide(color: Colors.blueAccent, width: 2.0),
+borderRadius: BorderRadius.all(Radius.circular(32.0)),
+),
+);
